+++ cmd/auth/auth.go
@@ -33,9 +33,6 @@ func NewCmdAuth(f *cmdutil.Factory) *cobra.Command {
 	cmd.AddCommand(NewCmdAuthScopes(f, nil))
 	cmd.AddCommand(NewCmdAuthList(f, nil))
 	cmd.AddCommand(NewCmdAuthCheck(f, nil))
-	if proxyCmd := NewCmdAuthProxy(f); proxyCmd != nil {
-		cmd.AddCommand(proxyCmd)
-	}
 	return cmd
 }
