+++ extension/credential/sidecar/provider_test.go
@@ -12,7 +12,7 @@ import (
 
 	"github.com/larksuite/cli/extension/credential"
 	"github.com/larksuite/cli/internal/envvars"
-	"github.com/larksuite/cli/internal/sidecar"
+	"github.com/larksuite/cli/sidecar"
 )
 
 func setEnv(t *testing.T, key, value string) {
