+++ extension/transport/sidecar/interceptor_test.go
@@ -12,7 +12,7 @@ import (
 	"net/http"
 	"testing"
 
-	"github.com/larksuite/cli/internal/sidecar"
+	"github.com/larksuite/cli/sidecar"
 )
 
 // failingBody is a ReadCloser that errors on Read and tracks Close calls.
