+++ sidecar/hmac.go
