+++ sidecar/hmac_test.go
